Removed advert. But Button is flickering. Remove ads screen with explanation etc.
Tournament Scheduler relies on your support to fund its development. The app is totally free for you to use without any limitation or reservation. If you find it useful, please consider leaving a donation in the form of advert removal. Yes, you are paying for advertisements to be removed. This one-off payment is less than the amount you would pay for couple of beers. So cheers to that!